add permessions to the system

+++ routes/web.php
@@ -3,6 +3,7 @@
 use App\Http\Controllers\ConstantController;
 use App\Http\Controllers\EngineerController;
 use App\Http\Controllers\HomeController;
+use App\Http\Controllers\UserController;
 use App\Http\Controllers\ProfileController;
 use Illuminate\Support\Facades\Route;
 
@@ -20,20 +21,72 @@ Route::get('/login', function () {
 Route::middleware(['auth', 'verified'])->group(function () {
 
     Route::get('/dashboard', [HomeController::class, 'index'])
-        ->name('dashboard');
+        ->name('dashboard')->middleware('permission:dashboard.view');
     Route::resource('constants', ConstantController::class);
 
+// عرض المستخدمين
+Route::get('users', [UserController::class, 'index'])
+    ->name('users.index')
+    ->middleware('permission:users.view');
+
+// صفحة إنشاء مستخدم
+Route::get('users/create', [UserController::class, 'create'])
+    ->name('users.create')
+    ->middleware('permission:users.create');
+
+// حفظ مستخدم جديد
+Route::post('users', [UserController::class, 'store'])
+    ->name('users.store')
+    ->middleware('permission:users.create');
+
+// صفحة تعديل مستخدم
+Route::get('users/{id}/edit', [UserController::class, 'edit'])
+    ->name('users.edit')
+    ->middleware('permission:users.edit');
+
+// تحديث بيانات مستخدم
+Route::put('users/{id}', [UserController::class, 'update'])
+    ->name('users.update')
+    ->middleware('permission:users.edit');
+
+// حذف مستخدم
+Route::delete('users/{id}', [UserController::class, 'destroy'])
+    ->name('users.destroy')
+    ->middleware('permission:users.delete');
+    
+Route::get('/users/{id}', [UserController::class, 'show'])->name('users.show');
+
+// جلب صلاحيات الدور عبر AJAX
+Route::get('users/get-role-permissions/{roleId}', [UserController::class, 'getRolePermissions'])
+    ->name('users.role.permissions')
+    ->middleware('permission:users.create'); 
+
     Route::prefix('engineers')->name('engineers.')->group(function () {
-        Route::get('/', [EngineerController::class, 'index'])->name('index');
-        Route::get('/create', [EngineerController::class, 'create'])->name('create');
-        Route::post('/', [EngineerController::class, 'store'])->name('store');
-        Route::get('/{engineer}', [EngineerController::class, 'show'])->name('show');
-        Route::get('/{engineer}/edit', [EngineerController::class, 'edit'])->name('edit');
-        Route::put('/{engineer}', [EngineerController::class, 'update'])->name('update');
-        Route::delete('/{engineer}', [EngineerController::class, 'destroy'])->name('destroy');
+        Route::get('/', [EngineerController::class, 'index'])->name('index')->middleware('permission:engineers.view');
+        Route::get('/create', [EngineerController::class, 'create'])->name('create')->middleware('permission:engineers.create');
+        Route::post('/', [EngineerController::class, 'store'])->name('store')->middleware('permission:engineers.create');
+        Route::get('/{engineer}', [EngineerController::class, 'show'])->name('show')->middleware('permission:engineers.view');
+        Route::get('/{engineer}/edit', [EngineerController::class, 'edit'])->name('edit')->middleware('permission:engineers.edit');
+        Route::put('/{engineer}', [EngineerController::class, 'update'])->name('update')->middleware('permission:engineers.edit');
+        Route::delete('/{engineer}', [EngineerController::class, 'destroy'])->name('destroy')->middleware('permission:engineers.delete');
         
         Route::get('/cities/{governorate}', [EngineerController::class, 'getCities'])->name('cities');
     });
+
+    Route::get('/profile', [ProfileController::class, 'index'])
+        ->name('profile.index');
+
+    Route::get('/profile/edit', [ProfileController::class, 'edit'])
+        ->name('profile.edit');
+
+    Route::post('/profile/update', [ProfileController::class, 'update'])
+        ->name('profile.update');
+
+    Route::get('/profile/change-password', [ProfileController::class, 'changePassword'])
+        ->name('profile.change-password');
+
+    Route::put('/profile/update-password', [ProfileController::class, 'updatePassword'])
+        ->name('profile.update-password');
 
 });
